structural movement of files
