+++ src/main/java/bot/MemeBot2000.java
@@ -1,4 +1,4 @@
-package app;
+package bot;
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.File;
