+++ src/main/java/datastructures/MemeBotMsg2000.java
@@ -1,4 +1,4 @@
-package dataStructures;
+package datastructures;
 
 import org.json.JSONObject;
