+++ src/main/java/datastructures/MemeDBMsg2000.java
@@ -1,4 +1,4 @@
-package dataStructures;
+package datastructures;
 
 import java.util.List;
