+++ src/main/java/db/MemeDB2000.java
@@ -1,4 +1,4 @@
-package app;
+package db;
 
 import java.sql.*;
 import java.util.ArrayList;
