+++ src/test/java/datastructures/LoggerTester.java
@@ -1,4 +1,4 @@
-package app;
+package datastructures;
 
 import java.io.FileNotFoundException;
 import java.util.LinkedList;
@@ -14,8 +14,8 @@ import org.junit.rules.TestName;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
-import dataStructures.Logger;
-import dataStructures.Logger.level;
+import datastructures.Logger;
+import datastructures.Logger.level;
 
 public class LoggerTester {
